requirements.txt
discord.py
python-dotenv
requests
bs4
colorama